+++ lib/sparql/algebra/translation.ex
@@ -507,7 +507,7 @@ defmodule SPARQL.Algebra.Translation do
   defp translate_graph_pattern(%OptionalGraphPattern{expr: expr}, state) do
     case translate_graph_pattern(expr, state) do
       %GroupGraphPattern{expr: expr, fs: []} ->
-        %SPARQL.Algebra.LeftJoin{expr2: expr, filters: RDF.true}
+        %SPARQL.Algebra.LeftJoin{expr2: expr, filters: [RDF.true]}
 
       %GroupGraphPattern{expr: expr, fs: fs} ->
         %SPARQL.Algebra.LeftJoin{expr2: expr, filters: fs}
@@ -516,7 +516,7 @@ defmodule SPARQL.Algebra.Translation do
 
       # TODO: remove this when the implementation is complete; we currently need this to make the W3C syntax tests pass on non-select queries
       nil ->
-       %SPARQL.Algebra.LeftJoin{expr2: @zero_bgp, filters: RDF.true}
+       %SPARQL.Algebra.LeftJoin{expr2: @zero_bgp, filters: [RDF.true]}
     end
   end